requirements.txt
Flask
pymongo
python-dotenv
psutil
requests